EI, JSTL 적용 및 replace와 historyback을 통한 예외처리

+++ src/main/java/exam/article/ArticleController.java
@@ -4,6 +4,7 @@ package exam.article;
 import exam.Rq;
 import exam.article.dto.ArticleDto;
 
+import java.io.IOException;
 import java.util.List;
 
 public class ArticleController {
@@ -26,44 +27,76 @@ public class ArticleController {
     }
 
     public void showDetail(Rq rq, Long id) {
+
         ArticleDto articleDto = articleService.findById(id);
+
+        if (articleDto == null) {
+            rq.historyBack("해당 글은 존재하지 않습니다.");
+            return;
+        }
 
         rq.setAttr("article", articleDto);
         rq.view("usr/article/detail");
     }
 
-    public void doWrite(Rq rq) {
+    public void doWrite(Rq rq) throws IOException {
         String title = rq.getParam("title", "");
         String body = rq.getParam("body", "");
 
+        if (title.length() == 0) {
+            rq.historyBack("제목을 입력해주세요.");
+            return;
+        }
+
+        if (body.length() == 0) {
+            rq.historyBack("내용을 입력해주세요.");
+            return;
+        }
+
         long id = articleService.write(title, body);
-        rq.appendBody("%d번 게시물이 생성 되었습니다.".formatted(id));
-        rq.appendBody("<br/>");
-        rq.appendBody("<a href=/usr/article/list/free>게시글 목록</a>");
+        rq.replace("/usr/article/detail/free/%d".formatted(id), "%d번 게시물이 생성 되었습니다.".formatted(id));
     }
 
     public void deleteArticle(Rq rq, Long id) {
+
+        ArticleDto articleDto = articleService.findById(id);
+
+        if (articleDto == null) {
+            rq.historyBack("해당 글은 존재하지 않습니다.");
+            return;
+        }
+
         articleService.deleteArticle(id);
 
-        rq.appendBody("%d번 게시물이 삭제되었습니다.".formatted(id));
-        rq.appendBody("<br/>");
-        rq.appendBody("<a href=/usr/article/list/free>게시글 목록</a>");
+        rq.replace("/usr/article/list/free", "%d번 게시물이 삭제되었습니다.".formatted(id));
     }
 
     public void showModifyForm(Rq rq, Long id) {
+
         ArticleDto articleDto = articleService.findById(id);
+
+        if (articleDto == null) {
+            rq.historyBack("해당 글이 존재하지 않습니다.");
+            return;
+        }
 
         rq.setAttr("article", articleDto);
         rq.view("usr/article/modify");
     }
 
     public void doModify(Rq rq, long id) {
+
+        ArticleDto articleDto = articleService.findById(id);
+
+        if (articleDto == null) {
+            rq.historyBack("해당 글이 존재하지 않습니다.");
+            return;
+        }
+
         String title = rq.getParam("title", "");
         String body = rq.getParam("body", "");
 
         articleService.modifyArticle(id, title, body);
-        rq.appendBody("%d번 게시물이 수정 되었습니다.".formatted(id));
-        rq.appendBody("<br/>");
-        rq.appendBody("<a href=/usr/article/list/free>게시글 목록</a>");
+        rq.replace("/usr/article/detail/free/%d".formatted(id), "%d번 게시물이 수정되었습니다.".formatted(id));
     }
 }
